% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_roc_rr.R
\name{add_ci}
\alias{add_ci}
\title{Adds lower and upper bounds of Confidence Interval to df}
\usage{
add_ci(df, ci.level, target)
}
\arguments{
\item{df}{data.frame containing tpr, fpr and their variances}

\item{ci.level}{double between 0-1.}

\item{target}{outcome vector

#' @return 
a data.frame with the elements:
\itemize{
  \item{tpr}{logit-transformed TPR values}
  \item{fpr}{logit-transformed FPR values}
  \item{var_tpr}{logit-transformed variance for TPR}
  \item{ll_tpr}{lower limit of CI for logit-transformed TPR values}
  \item{ul_tpr}{upper limit of CI for logit-transformed TPR values}
}}
}
\description{
Adds lower and upper bounds of Confidence Interval to df
}
\keyword{Internal}
