% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/rcppdocumentation.R
\name{apply_cut_off}
\alias{apply_cut_off}
\title{Apply all possible cutoffs}
\usage{
apply_cut_off(score, unique_vals)
}
\arguments{
\item{score}{the score}

\item{unique_vals}{the possible cutoffs}
}
\value{
A matrix of dimensions length(score) x length(cutoffs)

A matrix of dimensions length(score) x length(cutoffs)
}
\description{
Apply all possible cutoffs

Apply all possible cutoffs
}
\keyword{Internal}
