% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{apply_metrics}
\alias{apply_metrics}
\title{Function to apply metrics}
\usage{
apply_metrics(m, r)
}
\arguments{
\item{m}{matrix with one column for each cutoff and one row for each 
value in score}

\item{r}{binary response vector}
}
\value{
a list with 3 elements:
\item{tpri}{TPR values}
\item{fpri}{FPR values}
\item{vartpri}{variance for TPR}
}
\description{
Function to apply metrics
}
\keyword{Internal}
