% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_roc_rr.R
\name{backtransform_df}
\alias{backtransform_df}
\title{Adds backtransformed tpr, fpr and optionally confidence intervals to df}
\usage{
backtransform_df(df, ci)
}
\arguments{
\item{df}{data.frame containing tpr, fpr and their variances}

\item{ci}{Logical, is there a ci to backtransform

#' @return 
a data.frame with the elements:
\item{tpr_logit}{logit-transformed TPR values}
\item{fpr_logit}{logit-transformed FPR values}
\item{var_tpr_logit}{logit-transformed variance for TPR}
\item{tpr}{TPR}
\item{fpr}{FPR}
\item{ll_tpr_logit}{lower limit of CI for logit-transformed TPR values}
\item{ul_tpr_logit}{upper limit of CI for logit-transformed TPR values}
\item{ll_tpr}{lower limit of CI for TPR values}
\item{ul_tpr}{upper limit of CI for TPR values}}
}
\description{
Adds backtransformed tpr, fpr and optionally confidence intervals to df
}
\keyword{Internal}
