% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_roc_test.R
\name{cases_controls.long}
\alias{cases_controls.long}
\title{Extract cases and controls for two groups}
\usage{
cases_controls.long(d, target, score, group, groups, m)
}
\arguments{
\item{d}{a data.frame}

\item{target}{character, the outcome variable}

\item{score}{character, the prediction variable}

\item{group}{character, the group variable}

\item{groups}{vector with the two group levels to compare}

\item{m}{imputation number}
}
\description{
Helper function to prepare data for delongPlacements
}
\keyword{Internal}
