% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_roc_test.R
\name{cases_controls.wide}
\alias{cases_controls.wide}
\title{Extract cases and controls for two groups}
\usage{
cases_controls.wide(d, target, score, score2)
}
\arguments{
\item{d}{a data.frame}

\item{target}{character, the outcome variable}

\item{score}{character, the prediction variable for group 1}

\item{score2}{character, the prediction variable for group 2}
}
\description{
Helper function to prepare data for delongPlacements
}
\keyword{Internal}
