% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_roc_rr.R
\name{combine_metrics}
\alias{combine_metrics}
\title{Transforms a list of lists to a list of three matrices}
\usage{
combine_metrics(ll)
}
\arguments{
\item{ll}{A list of lists. For each imputation, one `l` as described in 
`transform_metrics`.

#' @return 
a list with 3 elements:
\item{tprm}{matrix of logit-transformed TPR values, one column for each 
imputation}
\item{fprm}{vector of logit-transformed FPR values passed on from `fpr_vals`, 
one column for each imputation}
\item{vartprm}{matrix of logit-transformed variance for TPR, one column for 
each imputation}}
}
\description{
Transforms a list of lists to a list of three matrices
}
\keyword{Internal}
