% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count_vals}
\alias{count_vals}
\title{Function to count number of val in x}
\usage{
count_vals(x, val)
}
\arguments{
\item{x}{IntegerVector}

\item{val}{int, the value to count}
}
\description{
Function to count number of val in x
}
\keyword{Internal}
