% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/rcppdocumentation.R
\name{cut_off}
\alias{cut_off}
\title{check if a value is bigger than a cutoff}
\usage{
cut_off(score, cutoff)
}
\arguments{
\item{score}{NumericVector of scores}

\item{cutoff}{double, the cutoff}
}
\description{
check if a value is bigger than a cutoff
}
\keyword{Internal}
