% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_roc}
\alias{get_roc}
\title{Calculates ROC curve (tpr and fpr at each cutoff)}
\usage{
get_roc(response, prediction, corr)
}
\arguments{
\item{response}{integer, response vector}

\item{prediction}{double, predictions}

\item{corr}{Continuity correction}
}
\description{
Using a fast greedy algorithm
}
\keyword{Internal}
