% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_roc_rr.R
\name{get_unique_vals}
\alias{get_unique_vals}
\title{Gets all unique values of a score in ascending order}
\usage{
get_unique_vals(data, score, digits = 3)
}
\arguments{
\item{data}{list of imputed data sets}

\item{score}{character, name of score variable in data}

\item{digits}{FALSE if values should not be rounded, otherwise any number 
of digits.}
}
\description{
Gets all unique values of a score in ascending order
}
\keyword{Internal}
