% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{order_descending}
\alias{order_descending}
\title{C++ implementation of R `order`
Assuming we don't have any missing values, only one vector to order and 
always want to order descending.}
\usage{
order_descending(x)
}
\arguments{
\item{x}{double vector to order}
}
\description{
C++ implementation of R `order`
Assuming we don't have any missing values, only one vector to order and 
always want to order descending.
}
\keyword{Internal}
