% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool-auc-rr.R
\name{pool_auc_rr}
\alias{pool_auc_rr}
\title{Pool AUC after Multiple Imputation}
\usage{
pool_auc_rr(rocs, ci.level = 0.95, transform = "logit")
}
\arguments{
\item{rocs}{A list of `pROC::roc` objects for each imputed data set}

\item{ci.level}{CI level for the AUC. Double between 0-1, by default 0.95.}

\item{transform}{A character string with the transformation if the AUC should 
be transformed before pooling or FALSE if the raw AUC should be pooled. 
"logit" by default.}
}
\description{
Pools AUC, it's variance and confidence intervals from a list of
roc objects using the methods described by Delong. 
This function assumes that we did NOT impute the dependent variable.
}
\section{Functions}{
\itemize{
\item \code{pool_auc_rr()}: 

}}
\author{
Jonas Schropp, code fragments from `pROC::roc.test` and `psfmi::pool_auc`.
}
