% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_roc_rr.R
\name{pool_roc_rr}
\alias{pool_roc_rr}
\title{Pool ROC curves based on multiple imputation}
\usage{
pool_roc_rr(
  data,
  score = "fib4",
  target = "advanced_fibrosis",
  unique_vals = NULL,
  fpr_vals = seq(from = 0.001, to = 0.999, by = 0.001),
  backtransform = FALSE,
  ci.level = NULL,
  tol = 0.001
)
}
\arguments{
\item{data}{A list of data.frames}

\item{score}{Character, the name of the score variable that should be present 
in each data.frame in data}

\item{target}{The name of the target variable present in each 
data.frame (must be binary).}

\item{unique_vals}{Optional. Define the unique values that should be used as 
cutoffs for the score variable. If not supplied, every unique cutoff is used.}

\item{fpr_vals}{The FPR values over which to pool the TPR values. Defaults 
to `seq(from=0.001, to=0.999, by=0.001)`}

\item{backtransform}{Should the results be transformed back?}

\item{ci.level}{Double between 0-1. NULL if no confidence intervals are 
desired (or you are going to bootstrap confidence intervals).}

\item{tol}{tolerance to use for 0 and 1 values (due to logit trans). 1e-3 
by default. `FALSE` if no tolerance should be used instead for 0 values.}
}
\value{
a data.frame with the elements:
\item{tpr_logit}{logit-transformed TPR values}
\item{fpr_logit}{logit-transformed FPR values}
\item{var_tpr_logit}{description a}
\item{tpr}{description a}
\item{fpr}{description a}
\item{ll_tpr_logit}{description a}
\item{ul_tpr_logit}{description a}
}
\description{
The function `pool_roc_rr` pools roc curves created from a list of multiply 
imputed data sets by applying Rubin's Rules to the logit-transformed 
TPR and FPR at each unique value of the supplied score. 
The function returns a data.frame of transformed TPR and FPR at each cutoff, 
their respective variances and optionally confidence intervals as well as 
backtransformed values.
}
