% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/rcppdocumentation.R
\name{pool_rr_m}
\alias{pool_rr_m}
\title{Pool mean and variance for many rows following Rubin's Rules}
\usage{
pool_rr_m(est, var)
}
\arguments{
\item{est}{a matrix of (logit-transformed) proportion by imp}

\item{var}{a matrix of (logit transformed) variance by imp}
}
\description{
Pool mean and variance for many rows following Rubin's Rules

Pool mean and variance for many rows following Rubin's Rules
}
\keyword{Internal}
