% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_roc_test.R
\name{prep_rocs.paired}
\alias{prep_rocs.paired}
\title{Prepares list of ROCs for paired case}
\usage{
prep_rocs.paired(data, target, score, score2, group, groups, levels, direction)
}
\arguments{
\item{data}{see mi_roc_test}

\item{target}{see mi_roc_test}

\item{score}{see mi_roc_test}

\item{score2}{see mi_roc_test}

\item{group}{see mi_roc_test}

\item{groups}{see mi_roc_test}

\item{levels}{see mi_roc_test}

\item{direction}{see mi_roc_test}
}
\description{
Prepares list of ROCs for paired case
}
\keyword{Internal}
