% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_roc_test.R
\name{prep_rocs.unpaired}
\alias{prep_rocs.unpaired}
\title{Prepares list of ROCs for unpaired case}
\usage{
prep_rocs.unpaired(data, target, score, group, groups, direction)
}
\arguments{
\item{data}{see mi_roc_test}

\item{target}{see mi_roc_test}

\item{score}{see mi_roc_test}

\item{group}{see mi_roc_test}

\item{groups}{see mi_roc_test}

\item{direction}{see mi_roc_test}
}
\description{
Prepares list of ROCs for unpaired case
}
\keyword{Internal}
