% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/rcppdocumentation.R
\name{rcpp_rowMeans}
\alias{rcpp_rowMeans}
\title{Equivalent to R rowMeans}
\usage{
rcpp_rowMeans(x)
}
\arguments{
\item{x}{a matrix}
}
\description{
Equivalent to R rowMeans

Equivalent to R rowMeans
}
\keyword{Internal}
