% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reorder}
\alias{reorder}
\title{Reorder a int vector by indices
equivalent to R vector[indices]}
\usage{
reorder(values, indices)
}
\arguments{
\item{values}{the vector to reorder/subset}

\item{indices}{integer vector of indices to order by}
}
\description{
Reorder a int vector by indices
equivalent to R vector[indices]
}
