% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/rcppdocumentation.R
\name{rowVars}
\alias{rowVars}
\title{Apply var to matrix rows}
\usage{
rowVars(x)
}
\arguments{
\item{x}{a matrix}
}
\description{
Apply var to matrix rows

Apply var to matrix rows
}
\keyword{Internal}
