% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tpr}
\alias{tpr}
\title{TPR}
\usage{
tpr(r, p, corr)
}
\arguments{
\item{r}{response vector, 0 / 1}

\item{p}{prediction vector, 0 / 1}

\item{corr}{continuity correction}
}
\description{
TPR
}
\keyword{Internal}
