% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_roc_rr.R
\name{transform_metrics}
\alias{transform_metrics}
\title{transform metrics}
\usage{
transform_metrics(l)
}
\arguments{
\item{l}{list containing 4 elements: a vector of `tpr`, a vector of `fpr` and 
the respective variances. The length of each vector is the number of 
distinct cutoffs (`unique_vals - 1`)

#' @return 
a list with 3 elements:
\itemize{
  \item{tpri}{logit-transformed TPR values}
  \item{fpri}{logit-transformed original FPR values}
  \item{vartpri}{logit-transformed variance for TPR}
}}
}
\description{
transform metrics
}
\keyword{Internal}
