% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/rcppdocumentation.R
\name{var_prop}
\alias{var_prop}
\title{Variance for a proportion}
\usage{
var_prop(est, n)
}
\arguments{
\item{est}{vector of tpr or fpr}

\item{n}{number of positives in r or negatives in r}
}
\description{
Variance for a proportion
}
\keyword{Internal}
