% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/rcppdocumentation.R
\name{varlogit}
\alias{varlogit}
\title{transform variances to logit scale}
\usage{
varlogit(var, est)
}
\arguments{
\item{var}{NumericVector of variances}

\item{est}{NumericVector of proportions}
}
\description{
transform variances to logit scale

transform variances to logit scale
}
\keyword{Internal}
