% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{varproplogit}
\alias{varproplogit}
\title{Variance for a logit-transformed proportion}
\usage{
varproplogit(events, n, corr)
}
\arguments{
\item{corr}{continuity correction}

\item{r}{response vector, 0 / 1}

\item{p}{prediction vector, 0 / 1}
}
\description{
Variance for a logit-transformed proportion
}
\keyword{Internal}
