% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_roc_rr.R
\name{add_var_roc}
\alias{add_var_roc}
\title{Adds roc variance to df following the formula in Martinez}
\usage{
add_var_roc(df, r)
}
\arguments{
\item{df}{data.frame with tpr, fpr, vartpr}

\item{r}{response vector}
}
\description{
varroc = vartpr + varfpr
}
\keyword{Internal}
