% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fpr}
\alias{fpr}
\title{FPR}
\usage{
fpr(r, p, corr)
}
\arguments{
\item{r}{response vector, 0 / 1}

\item{p}{prediction vector, 0 / 1}

\item{corr}{continuity correction}
}
\description{
FPR
}
\keyword{Internal}
