% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_roc_test.R
\name{mi_roc_test}
\alias{mi_roc_test}
\alias{mi_roc_test.unpaired}
\alias{mi_roc_test.paired}
\title{Pool Delong Test after Multiple Imputation}
\usage{
mi_roc_test(rocs1, rocs2, paired = FALSE)

mi_roc_test.unpaired(rocs1, rocs2)

mi_roc_test.paired(rocs1, rocs2)
}
\arguments{
\item{rocs1}{A list of "roc" objects for each imputed data set}

\item{rocs2}{A list of "roc" objects for each imputed data set 
to compare to rocs1}

\item{paired}{Are the roc curves paired?}
}
\description{
Pools the variance and covariance of two (un-)paired ROC after multiple 
imputation to compare their AUC. This function assumes that we did NOT 
impute the dependent variable.
}
\details{
Unpaired version

Paired version
}
\section{Functions}{
\itemize{
\item \code{mi_roc_test()}: 

\item \code{mi_roc_test.unpaired()}: 

\item \code{mi_roc_test.paired()}: 

}}
\author{
Jonas Schropp, code fragments from `pROC::roc.test` and `MKmisc::mi.t.test`.
}
\keyword{Internal}
