% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{varproplogit}
\alias{varproplogit}
\title{Variance for a logit-transformed proportion}
\usage{
varproplogit(events, n, corr)
}
\arguments{
\item{events}{double, number of events}

\item{n}{double, number of trials}

\item{corr}{continuity correction}
}
\description{
Variance for a logit-transformed proportion
}
\keyword{Internal}
